Extract GetClientSet to nativekubernetes
