+++ pkg/kubernetesutils/nativekubernetes/README.md
@@ -1,0 +1,5 @@
+# nativekubernetes package
+
+Non object oriented implementation using the [official client-go library](https://github.com/kubernetes/client-go) to interact with kubernetes.
+
+For the object oriented implmenetation see [nativekubernetesoo](/pkg/kubernetesutils/nativekubernetesoo/).
