+++ pkg/kubernetesutils/nativekubernetesoo/README.md
@@ -1,3 +1,5 @@
 # nativekubernetesoo package
 
 Object oriented implementation using the [official client-go library](https://github.com/kubernetes/client-go) to interact with kubernetes.
+
+For the non object oriented implmenetation see [nativekubernetes](/pkg/kubernetesutils/nativekubernetes/).
